Minor cleaning and renaming
